+++ src/Traits/Precisionable.php
@@ -3,14 +3,13 @@
 namespace Weboccult\EatcardCompanion\Traits;
 
 use Illuminate\Support\Facades\Log;
+use function Weboccult\EatcardCompanion\Helpers\{splitDigits};
 
 /**
  * @description To handle precision for database fields
  */
 trait Precisionable
 {
-	use Splitable;
-
 	public static function bootPrecisionable()
 	{
 		static::saving(function ($model) {
@@ -38,24 +37,24 @@ trait Precisionable
 								$precision = (int)$precisions['precision'];
 							}
 							//skip save difference calc and fixe split with two digits for below fields for cash payment
-							if (isset($model->method) && $model->method == 'cash' && in_array($field, $skip_reguler_split_fields)) {
+							if (isset($model->method) && $model->method == 'cash' && isset($model->payment_split_type) && empty($model->payment_split_type) && in_array($field, $skip_reguler_split_fields)) {
 								$skip_reguler_split = true;
 								$precision = 2;
 							}
 							//Check for any specific difference_field is declared or not | skip for cash payment methods
 							if (!$skip_reguler_split && isset($precisions['difference']) && !empty($precisions['difference'])) {
 								$difference_field = $precisions['difference'];
-								if ($model->hasAttribute($difference_field) && in_array($difference_field, $model->fillable)) {
+								if (/*$model->hasAttribute($difference_field) && */ in_array($difference_field, $model->fillable)) {
 									if (isset($precisions['difference_precision']) && !empty($precisions['difference_precision'])) {
 										$difference_precision = (int)$precisions['difference_precision'];
 									}
 									if (isset($precisions['payment_digit']) && !empty($precisions['payment_digit'])) {
 										$payment_digit = (int)$precisions['payment_digit'];
 									}
-									$model->$difference_field = $model->splitDigits($model->$field, $precision, true, $difference_precision, $payment_digit);
+									$model->$difference_field = splitDigits($model->$field, $precision, true, $difference_precision, $payment_digit);
 								}
 							}
-							$model->$field = $model->splitDigits($model->$field, $precision);
+							$model->$field = splitDigits($model->$field, $precision);
 						}
 					}
 				}