+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to `eatcard-companion` will be documented in this file
 
-## 0.0.1 - 2021-11-24
+## 1.1 - 2021-12-08
+
+- Improvements added in Precisionable trait.
+- Splitable trait removed in favour of new splitDigits function.
+
+## 1.0 - 2021-11-24
 
 - initial release